fix(georanges,classifier): close both paths that route Russian addresses through the tunnel (#270)
Russian addresses kept reaching the tunnel by two independent routes,
and the existing exclusion closed neither completely.

The data gap. georanges-gen sourced country attribution from the RIR
delegated-statistics files alone, which record the country of the
top-level allocation and never restate it when a block is sub-assigned
abroad. The row covering the incident address 138.124.255.106 reads
"ripencc|CH|ipv4|138.124.240.0|4096|19930901|assigned" -- Switzerland --
while RIPE's own database puts country: RU on that /24, and the
surrounding legacy /17 is split between 36 ASNs, 7 of them Russian,
across 82 /24s. The shipped list contained nothing at all in
138.124.0.0/16, so the veto never fired.

Fixed by adding a second layer built from data the command already
downloads: those same delegated-stats files record that AS214891, the
origin of that /24, is RU, so the only missing piece was a
prefix-to-origin-ASN map. iptoasn.com supplies it -- hourly, 7 MB,
public domain (PDDL v1.0), preferred over Loyalsoldier/geoip
(CC-BY-SA-4.0, daily) and bgp.tools (76 MB, non-commercial). A row is
accepted only when both sources agree the ASN is Russian; on a real run
all 26855 rows agreed and none were rejected. Both layers are then
coalesced into a minimal CIDR set, which makes the file smaller (10508
lines, was 11450) while covering 1.2M more addresses (46.42M, was
45.22M). A minAddresses guard joins minRanges, since coalescing means a
healthy line count no longer implies healthy coverage.

The structural gap. isExcludedDst gates classification, per flow -- but
ClrBlockPromote does not classify. It widens a threshold-crossing block
into one ipset entry, and the REDIRECT rule then matches every address
inside it, so a Russian address in a promoted neighbour's /24 (or, via a
known-range match, its /18) rode along with the veto never consulted.
Mixed legacy space makes that ordinary rather than exotic.

Fixed with a block-level counterpart, georanges.Overlaps wired in as
classifier.Config.ExcludedRangeOverlap. A tainted known-range match
narrows back to the naive block first, and only a tainted naive block
abandons widening entirely -- the confirmed addresses keep the
individual promotions they already earned. Widening is an optimization;
redirecting Russian traffic is a defect.

georanges-update.yml moves weekly -> daily: the old cadence rested on
registry delegations changing slowly, which no longer describes the ASN
layer.

Verified by removing the ClrBlockPromote guard and confirming the new
tests fail with the old behaviour (1.2.3.0/24 and 1.2.0.0/18 promoted),
and by a real generator run plus an embedded-data test asserting
138.124.255.106 now matches 138.124.254.0/23.

Not closed by this change: coverage is still not total (a Russian
service on a foreign ASN or behind a foreign CDN appears in neither
layer), and everything here is IPv4, as the whole adaptive-routing
dataplane is.

+++ cmd/georanges-gen/main_test.go
@@ -54,7 +54,7 @@ func TestCidrsForRange_SingleAddress(t *testing.T) {
 // actually matters for the veto: over-claiming would exclude somebody
 // else's addresses from the tunnel.
 func TestCidrsForRange_NeverCoversMoreThanAsked(t *testing.T) {
-	for _, count := range []uint32{1, 3, 5, 100, 255, 256, 1000, 1536, 4097, 65535} {
+	for _, count := range []uint64{1, 3, 5, 100, 255, 256, 1000, 1536, 4097, 65535} {
 		got := renderNets(cidrsForRange(ipv4("172.16.3.0"), count))
 		if total := addresses(got); total != int(count) {
 			t.Errorf("count %d: blocks cover %d addresses (%v)", count, total, got)
